Finer-grained selection of neighboring vertices in a voronoi graph
This allows very skewed grids to be detected. Prior to this patch I would only
consider vertices as neighbors if they were directly connected in the voronoi
diagram. In this diagram:

E--B-----D
| / \   /
|/   \ /
A-----C

I would consider E,B,C to be neighbors of A. I now also check the midpoints. So
I also consider D as a neighbor since it's directly between B and C